requirements.txt
openai==2.1.0
pandas==2.0.0
pdfplumber==0.10.0
PyPDF2==3.0.0
tabula-py==2.8.0
openpyxl==3.1.0
python-dotenv==1.1.1
grok==6.1